Move AddPostPage and BlogPage to post folder
